int-as-pointer parameters: re-type 22 whole components atomically (byte-neutral, 1009/1009)
The class C97 measured and declined: a parameter declared `int`/`u32` that the
body immediately casts to a pointer.  ABI-identical, so no gate in the tree can
see it; it is pure semantic debt.  C97's reason for declining was that the shape
is interlocked -- re-typing one end only relocates the cast -- so this lands the
CONNECTED COMPONENT of each parameter (definition + every prototype + every call
site + the interface function-pointer typedef where the DLL export table carries
it) in one edit, and the cast disappears from both ends.

22 components, 30 (function, parameter) slots:

  camera interface  Camera_setTarget / Camera_setTargetReticleOverride, both
                    CameraSetTargetFn / CameraSetTargetReticleOverrideFn and the
                    5 call sites through (*gCameraInterface); the 6 sites that
                    already pass 0 need no edit.  camera_interface.h picks up
                    game/objects/object_fwd.h (a pure forward typedef, emits
                    nothing into an allocated section).
  GameObject*       Lightfoot_ProcessHitResponseFlags, Lightfoot_UpdatePlayerInteraction,
                    cMenuPlayTrickyCommandSfx, cfguardian_playEventSfx,
                    kytesmum_playAnimationEventSfx, ring_checkArwingCollision,
                    dimmagicbridge_scrollTextureChannels, TrickyCurve_updateCooldownTrigger
  ExplodableChunk*  explodable_spawnFragmentObject, explodable_computeFragmentLaunch
                    (the redundant `chunk = chunkAddress` copies go with them)
  Texture*          setupWaterReflectionTev x2, setupReflectionDistortTev
  u8*               modelLoad_layoutBuffers, ObjModel_LoadAnimData
  int*              trackBuildModelTriangles
  void*             SH_LevelControl_doThornTailEvents, arwarwing_emitDamageEffects,
                    DIMSnowHorn1_spawnFootstepEffects, snowPrintSnowCloud /
                    newclouds_renderSnowClouds, vfplift_trigger

Gate: full rebuild of all 1009 objects compared against a pristine control built
at the same parent -- 1009 compared / 0 differ / 0 missing / 0 new (3 units carry
only @NNN literal-pool renumbering at preserved addresses, which never reaches
the link).  Tree unchanged at 99.81738 / matched_code 2519540 / matched_data
1204217 / complete_units 917 of 1046; forced link main.dol OK, sha1
e750e8e894707a52446118a4b84f1b58b677b269.

NOT byte-neutral, measured and reverted: staff_setupSwipe#3 (226 99.91625 ->
99.84835) and dfshShrine_updateHoverMotion#0 (DFSH_Shrine 100.0 -> 99.7479).  In
both, the body opens `GameObject* obj = (GameObject*)objArg;`, and once the
parameter is already a pointer MWCC coalesces the copy into one web and
re-colours the function.  The int spelling is load-bearing there.

+++ src/dlls/engine/19/19.c
@@ -535,7 +535,7 @@ void waterfx_render(int obj, int renderParam)
         j = 0;
         if (gWaterfxSplashCount != 0)
         {
-            setupWaterReflectionTev((int)gWaterfxSplashTexture0, (int)gWaterfxSplashTexture1);
+            setupWaterReflectionTev(gWaterfxSplashTexture0, gWaterfxSplashTexture1);
             GXSetArray(GX_VA_POS, gWaterfxSplashPosArray, 0xc);
             GXSetArray(GX_VA_TEX0, gWaterfxSplashTexCoordArray, 8);
             GXClearVtxDesc();
@@ -574,7 +574,7 @@ void waterfx_render(int obj, int renderParam)
         }
         if (gWaterfxWakeCount != 0)
         {
-            setupReflectionDistortTev((int)gWaterfxWakeTexture);
+            setupReflectionDistortTev(gWaterfxWakeTexture);
         }
         for (poolOffset = 0, j = 0, descriptorOffset = 0, vertexOffset = 0;
              j < WATERFX_POOL_SIZE;
